requirements.txt
matplotlib==3.6.2
numpy==1.23.5
numpy-stl==2.17.1
